weatherEats
WeatherEats is a native iOS app that displays all the top-rated
restaurants in the area based on a weather filter chosen by users. The
map also displays the users current location, in addition to all the
top restaurants they are around. This feature makes it convenient for
users to choose a top restaurant destination to their liking.